requirements.txt
google-cloud-storage==1.14.0
Flask==1.0.2
